api/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
av==10.0.0
blosc2==2.0.0
celluloid==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.1
deepdish==0.3.7
easing-functions==1.0.4
exceptiongroup==1.1.3
fastapi==0.102.0
filelock==3.12.2
fonttools==4.42.1
h11==0.14.0
h5py==3.9.0
httptools==0.6.0
idna==3.4
imageio==2.31.1
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
lit==16.0.6
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
nibabel==5.1.0
nilearn==0.10.1
nltools==0.4.7
numexpr==2.8.4
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
py-cpuinfo==9.0.0
py-feat==0.6.1
pydantic==2.3.0
pydantic_core==2.6.3
pynv==0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tables==3.8.0
threadpoolctl==3.2.0
tifffile==2023.8.25
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
xgboost==1.7.6
zipp==3.16.2